Move all code wrapping get_* and do_* into BaseJavascriptModule
All module types with custom JS will need these wrappers, including
in particular builtin modules.
